Pakistan-Legal-AI/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-groq
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
pymupdf
fpdf2
